+++ src/main/java/com/example/demo/task/CDPublishReadyTask.java
@@ -40,8 +40,8 @@ public class CDPublishReadyTask extends SimpleTask implements Runnable{
         String mcu_id = jsonObject.getString("mcu_id");
         String mcu_domain = jsonObject.getString("mcu_domain");
         String room_id = jsonObject.getString("room_id");
-        Integer audio_ssrc = jsonObject.getInteger("audio_ssrc");
-        Integer video_ssrc = jsonObject.getInteger("video_ssrc");
+        Long audio_ssrc = jsonObject.getLong("audio_ssrc");
+        Long video_ssrc = jsonObject.getLong("video_ssrc");
         JSONObject options_msg = jsonObject.getJSONObject("options");
         if(client_id==null||stream_id==null||mcu_id==null||mcu_domain==null||room_id==null||options_msg==null){
             log.error("{} msg lack params, msg: {}", CDPublishReadyTask.taskType, msg);
